requirements.txt
docx2txt
python-multipart
fastapi
uvicorn
langchain
langchain-chroma
langchain-community
langchain-core
langchain-openai
langgraph
openai
pydantic
pypdf
python-docx
python-dotenv
regex
requests
spacy
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl
streamlit
streamlit-chat
